fix: shared Chromium browser pool to prevent OOM kills during news collection
Previously every renderJavaScriptPage() call launched a new Chromium process
(~400MB each). With MAX_CONCURRENCY=10, up to 10 simultaneous Chromium
instances pushed the container past its 4g memory limit, causing OOM kills
that crashed the news collection job mid-run.

- Extract shared browser pool (acquireBrowser/releaseBrowser) into browserPool.js
- contentExtractor.js and jsRenderer.js both import from the shared pool
- jsRenderer hard-timeout now closes the BrowserContext, not the browser process
- Reduce teamSize 5→3 for per-POI pg-boss worker queue
- Fix local dev run.sh to use --network=host for IPv4+IPv6 localhost support

Memory profile: 1 shared Chromium (~250MB) + N contexts (~80MB each)
vs prior: N × Chromium (~400MB each). Peak drops from ~4GB to ~1GB at concurrency=10.
